DAW/UD8/Servidor/Proyecto/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2
django-allauth==0.59.0
django-crispy-forms==2.1
idna==3.6
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
psycopg2-binary==2.9.9
django-filter==23.5
djangorestframework==3.14.0
Markdown==3.1.1

